[refactor] Remove resources.go
Move most functions to sprite_sheet.go as they are related to sprite-sheet.
